backend/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.3
#3.2.12
# django-debug-toolbar==3.2.4
django-filter==21.1
# django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
#3.12.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==4.0.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
# isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.6.1
oauthlib==3.2.0
openapi-codec==1.3.2
Pillow==9.0.1
pycodestyle==2.8.0
psycopg2-binary==2.9.3
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.6
requests==2.27.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
# xhtml2pdf==0.2.5
